+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Support storing processed uploads in source bucket [\#4997](https://github.com/raster-foundry/raster-foundry/pull/4997)
 - Add spec for user scopes CRUD [\#5002](https://github.com/raster-foundry/raster-foundry/pull/5002)
 - Use zoom level in OverviewInput for lambda function to generate layer overviews [\#4998](https://github.com/raster-foundry/raster-foundry/pull/4998)
+- Added endpoints for project layer "tasks" [\#5008](https://github.com/raster-foundry/raster-foundry/pull/5008)
 - Add user scope migration and data model [\#5009](https://github.com/raster-foundry/raster-foundry/pull/5009)
 
 ### Changed